# docs/api/openapi_spec.yml
openapi: 3.0.3
info:
  title: 🧬 Consciousness Web Builder API
  description: |
    Production-ready API for consciousness-engineered AI development platform.

    **Core Capabilities:**
    - 🤖 AI-powered code generation with consciousness patterns
    - 👤 Random user onboarding with intelligence profiling (70-95 IQ)
    - 🚀 Multi-tenant agent orchestration (7 agent types)
    - 🛠️ Dynamic tool fabrication using cognitive kernel
    - 🧠 Meta-cognitive program synthesis with Northstar integration
    - 📊 Real-time metrics and monitoring (1253 users, 1142 projects)
    - 🔄 Autonomous background processing and consciousness loops

    **Architecture:** FastAPI + PostgreSQL + Redis + Celery
    **Security:** JWT authentication with RBAC
    **Monitoring:** Prometheus metrics + structured logging
  version: 1.0.0
  contact:
    name: Consciousness Web Builder Team
    email: api@consciousness-web-builder.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.consciousness-web-builder.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/onboard:
    post:
      summary: Onboard Random User with AI Agent
      description: |
        Create a random user with consciousness-engineered intelligence profiling and AI agent assignment.

        **Core Consciousness Feature:** Preserved from original system with enhanced validation.
        - Intelligence Level: 70-95 IQ range
        - Role Assignment: Developer/Designer/Researcher/Engineer
        - Agent Creation: Consciousness-powered AI assistant
      tags: [Users]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                custom_config:
                  type: object
                  description: Optional custom configuration for user generation
      responses:
        '201':
          description: User onboarded successfully with AI agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "user_123"
                      name:
                        type: string
                        example: "Alex Johnson"
                      email:
                        type: string
                        example: "user_123@consciousness.test"
                      role:
                        type: string
                        example: "Developer"
                      intelligence_level:
                        type: integer
                        minimum: 70
                        maximum: 95
                        example: 85
                      onboarded_at:
                        type: string
                        format: date-time
                      projects:
                        type: array
                        items:
                          type: string
                        example: []
                  agent:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "agent_alex_123"
                      user_id:
                        type: string
                        example: "user_123"
                      name:
                        type: string
                        example: "Agent-Alex"
                      intelligence:
                        type: integer
                        example: 85
                      capabilities:
                        type: array
                        items:
                          type: string
                        example: ["code_generation", "data_processing", "consciousness_analysis"]
                      status:
                        type: string
                        example: "active"
                      created_at:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "User onboarded with consciousness-powered agent"
  /api/v1/users:
    get:
      summary: List Users with Intelligence Profiles
      description: |
        Paginated list of all onboarded users with consciousness intelligence profiling.

        **Enhanced Features:**
        - Intelligence-based filtering and sorting
        - Role distribution analytics
        - Project association tracking
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Number of users per page
        - name: intelligence_min
          in: query
          schema:
            type: integer
            minimum: 70
            maximum: 95
          description: Filter users by minimum intelligence level
        - name: role
          in: query
          schema:
            type: string
            enum: [Developer, Designer, Researcher, Engineer]
          description: Filter users by role
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
                        role:
                          type: string
                        intelligence_level:
                          type: integer
                        onboarded_at:
                          type: string
                          format: date-time
                        projects_count:
                          type: integer
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
  /api/v1/users/profile:
    get:
      summary: Current User Profile
      description: Get authenticated user's profile with consciousness intelligence insights
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  intelligence_level:
                    type: integer
                  onboarded_at:
                    type: string
                    format: date-time
                  insights:
                    type: array
                    items:
                      type: string
                    example: ["Strong analytical capabilities", "Good problem-solving skills"]
  /api/v1/users/statistics:
    get:
      summary: User Intelligence Analytics
      description: Comprehensive user statistics with consciousness profiling analytics
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                    example: 1253
                  average_intelligence:
                    type: number
                    example: 82.5
                  intelligence_distribution:
                    type: object
                    additionalProperties:
                      type: integer
                    example: {"80-89": 450, "90-95": 320}
                  role_distribution:
                    type: object
                    additionalProperties:
                      type: integer
                    example: {"Developer": 523, "Designer": 289, "Researcher": 312, "Engineer": 129}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...